+++ src/mjlab/asset_zoo/robots/__init__.py
@@ -1,3 +1,9 @@
+from mjlab.asset_zoo.robots.bpx.bpx_constants import (
+  BPX_ACTION_SCALE as BPX_ACTION_SCALE,
+)
+from mjlab.asset_zoo.robots.bpx.bpx_constants import (
+  get_bpx_robot_cfg as get_bpx_robot_cfg,
+)
 from mjlab.asset_zoo.robots.i2rt_yam.yam_constants import (
   YAM_ACTION_SCALE as YAM_ACTION_SCALE,
 )