+++ src/mjlab/asset_zoo/robots/bpx/__init__.py
@@ -1,0 +1,2 @@
+"""Black Panther X robot assets."""
+