+++ src/mjlab/tasks/tracking/config/bpx/__init__.py
@@ -1,0 +1,21 @@
+from mjlab.tasks.registry import register_mjlab_task
+from mjlab.tasks.tracking.rl import MotionTrackingOnPolicyRunner
+
+from .env_cfgs import bpx_flat_tracking_env_cfg
+from .rl_cfg import bpx_tracking_ppo_runner_cfg
+
+register_mjlab_task(
+  task_id="Mjlab-Tracking-Flat-BPX",
+  env_cfg=bpx_flat_tracking_env_cfg(),
+  play_env_cfg=bpx_flat_tracking_env_cfg(play=True),
+  rl_cfg=bpx_tracking_ppo_runner_cfg(),
+  runner_cls=MotionTrackingOnPolicyRunner,
+)
+
+register_mjlab_task(
+  task_id="Mjlab-Tracking-Flat-BPX-No-State-Estimation",
+  env_cfg=bpx_flat_tracking_env_cfg(has_state_estimation=False),
+  play_env_cfg=bpx_flat_tracking_env_cfg(has_state_estimation=False, play=True),
+  rl_cfg=bpx_tracking_ppo_runner_cfg(),
+  runner_cls=MotionTrackingOnPolicyRunner,
+)